🦄 refactor: CategoryController
