+++ api/routes/index.js
@@ -3,9 +3,11 @@ const router = express.Router();
 
 const AuthRouter = require('./AuthRouter');
 const NoteRouter = require('./NoteRouter');
+const CategoryRouter = require('./CategoryRouter');
 
 router.use('/auth', AuthRouter);
 router.use('/note', NoteRouter);
+router.use('/category', CategoryRouter);
 
 router.all('*', (req, res) => {
   return res.status(404).json({
